pyproject.toml
[project]
name = "nemoflow"
version = "0.1.0"
description = "Reliability oracle for AI agents"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "redis>=5.2.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.0.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "stripe>=12.0.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "fakeredis>=2.26.0",
    "aiosqlite>=0.20.0",
]
mcp = [
    "mcp>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
